# frontend/src/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "LifeGrid API",
  "version": "0.1.0"
 },
 "paths": {
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoginOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/totp": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Totp",
    "operationId": "totp_api_v1_auth_totp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TotpIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/otp/request": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Otp Request",
    "operationId": "otp_request_api_v1_auth_otp_request_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpRequestIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OtpSent"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/otp/verify": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Otp Verify",
    "operationId": "otp_verify_api_v1_auth_otp_verify_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpVerifyIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_v1_admin_users_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserOut"
         },
         "type": "array",
         "title": "Response List Users Api V1 Admin Users Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create User",
    "operationId": "create_user_api_v1_admin_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/users/{user_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user_api_v1_admin_users__user_id__patch",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/users/{user_id}/totp": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Enrol Totp",
    "description": "Issue a new authenticator secret. The URI is shown once; only the encrypted secret is stored.",
    "operationId": "enrol_totp_api_v1_admin_users__user_id__totp_post",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TotpEnrolment"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/settings/{key}": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Get Setting",
    "operationId": "get_setting_api_v1_admin_settings__key__get",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SettingOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Put Setting",
    "operationId": "put_setting_api_v1_admin_settings__key__put",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SettingIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SettingOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/compatibility": {
   "get": {
    "tags": [
     "compatibility"
    ],
    "summary": "Compatible",
    "operationId": "compatible_api_v1_compatibility_get",
    "parameters": [
     {
      "name": "component",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Component"
      }
     },
     {
      "name": "abo",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Abo"
      }
     },
     {
      "name": "rh",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Rh"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DonorGroup"
         },
         "title": "Response Compatible Api V1 Compatibility Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/compat-rules": {
   "get": {
    "tags": [
     "compatibility"
    ],
    "summary": "List Rules",
    "operationId": "list_rules_api_v1_admin_compat_rules_get",
    "parameters": [
     {
      "name": "component",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Component"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RuleRow"
         },
         "title": "Response List Rules Api V1 Admin Compat Rules Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "compatibility"
    ],
    "summary": "Replace Rules",
    "operationId": "replace_rules_api_v1_admin_compat_rules_put",
    "parameters": [
     {
      "name": "component",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Component"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "array",
        "items": {
         "$ref": "#/components/schemas/RuleRow"
        },
        "title": "Rows"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RuleRow"
         },
         "title": "Response Replace Rules Api V1 Admin Compat Rules Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Abo": {
    "type": "string",
    "enum": [
     "O",
     "A",
     "B",
     "AB"
    ],
    "title": "Abo"
   },
...
   "Component": {
    "type": "string",
    "enum": [
     "red_cells",
     "platelets",
     "plasma"
    ],
    "title": "Component"
   },
...
   "DonorGroup": {
    "properties": {
     "donor_abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "donor_rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "rank": {
      "type": "integer",
      "title": "Rank"
     }
    },
    "type": "object",
    "required": [
     "donor_abo",
     "donor_rh",
     "rank"
    ],
    "title": "DonorGroup"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginIn": {
    "properties": {
     "email": {
      "type": "string",
      "maxLength": 254,
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 512,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginIn"
   },
   "LoginOut": {
    "properties": {
     "totp_required": {
      "type": "boolean",
      "title": "Totp Required"
     },
     "challenge": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Challenge"
     },
     "session": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TokenOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "totp_required"
    ],
    "title": "LoginOut"
   },
...
   "Me": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "site_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Site Ids"
     }
    },
    "type": "object",
    "required": [
     "id",
     "role",
     "email",
     "site_ids"
    ],
    "title": "Me"
   },
...
   "OtpRequestIn": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 32,
      "title": "Phone"
     }
    },
    "type": "object",
    "required": [
     "phone"
    ],
    "title": "OtpRequestIn"
   },
   "OtpSent": {
    "properties": {
     "sent": {
      "type": "boolean",
      "title": "Sent"
     },
     "dev_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dev Code"
     }
    },
    "type": "object",
    "required": [
     "sent"
    ],
    "title": "OtpSent"
   },
   "OtpVerifyIn": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 32,
      "title": "Phone"
     },
     "code": {
      "type": "string",
      "pattern": "^\\d{6}$",
      "title": "Code"
     },
     "role": {
      "type": "string",
      "enum": [
       "donor",
       "requester"
      ],
      "title": "Role",
      "default": "donor"
     }
    },
    "type": "object",
    "required": [
     "phone",
     "code"
    ],
    "title": "OtpVerifyIn"
   },
...
   "Rh": {
    "type": "string",
    "enum": [
     "pos",
     "neg"
    ],
    "title": "Rh"
   },
   "Role": {
    "type": "string",
    "enum": [
     "bank_manager",
     "hospital_lead",
     "donor_coordinator",
     "donor",
     "requester",
     "admin",
     "auditor"
    ],
    "title": "Role"
   },
   "RuleRow": {
    "properties": {
     "recipient_abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "recipient_rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "donor_abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "donor_rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "rank": {
      "type": "integer",
      "maximum": 16.0,
      "minimum": 1.0,
      "title": "Rank"
     }
    },
    "type": "object",
    "required": [
     "recipient_abo",
     "recipient_rh",
     "donor_abo",
     "donor_rh",
     "rank"
    ],
    "title": "RuleRow"
   },
...
   "SettingIn": {
    "properties": {
     "value": {
      "title": "Value"
     }
    },
    "type": "object",
    "required": [
     "value"
    ],
    "title": "SettingIn"
   },
   "SettingOut": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "value": {
      "title": "Value"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "updated_by": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated By"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "key",
     "value",
     "version",
     "updated_by",
     "updated_at"
    ],
    "title": "SettingOut"
   },
...
   "TokenOut": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "const": "bearer",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "user": {
      "$ref": "#/components/schemas/Me"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "TokenOut"
   },
   "TotpEnrolment": {
    "properties": {
     "otpauth_uri": {
      "type": "string",
      "title": "Otpauth Uri"
     }
    },
    "type": "object",
    "required": [
     "otpauth_uri"
    ],
    "title": "TotpEnrolment"
   },
   "TotpIn": {
    "properties": {
     "challenge": {
      "type": "string",
      "title": "Challenge"
     },
     "code": {
      "type": "string",
      "pattern": "^\\d{6}$",
      "title": "Code"
     }
    },
    "type": "object",
    "required": [
     "challenge",
     "code"
    ],
    "title": "TotpIn"
   },
...
   "UserIn": {
    "properties": {
     "email": {
      "type": "string",
      "maxLength": 254,
      "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "site_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Site Ids",
      "default": []
     },
     "password": {
      "type": "string",
      "maxLength": 512,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "role",
     "password"
    ],
    "title": "UserIn"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "site_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Site Ids"
     },
     "active": {
      "type": "boolean",
      "title": "Active"
     },
     "totp": {
      "type": "boolean",
      "title": "Totp"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "role",
     "site_ids",
     "active",
     "totp"
    ],
    "title": "UserOut"
   },
   "UserPatch": {
    "properties": {
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Role"
       },
       {
        "type": "null"
       }
      ]
     },
     "site_ids": {
      "anyOf": [
       {
        "items": {
         "type": "string",
         "format": "uuid"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Site Ids"
     },
     "active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Active"
     }
    },
    "type": "object",
    "title": "UserPatch"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}